+++ src/Symfony/Bridge/Twig/Extension/FormExtension.php
@@ -23,25 +23,26 @@ use Symfony\Component\Form\ChoiceList\View\ChoiceView;
  */
 class FormExtension extends \Twig_Extension implements \Twig_Extension_InitRuntimeInterface
 {
-    /**
-     * This property is public so that it can be accessed directly from compiled
-     * templates without having to call a getter, which slightly decreases performance.
-     *
-     * @var TwigRendererInterface
-     */
-    public $renderer;
+    private $renderer;
 
-    public function __construct(TwigRendererInterface $renderer)
+    public function __construct(TwigRendererInterface $renderer = null)
     {
+        if (null !== $this->renderer) {
+            @trigger_error(sprintf('Passing a Twig Form Renderer to the "%s" constructor is deprecated since version 3.2 and won\'t be possible in 4.0. Pass the Twig_Environment to the TwigRendererEngine constructor instead.', static::class), E_USER_DEPRECATED);
+        }
         $this->renderer = $renderer;
     }
 
     /**
      * {@inheritdoc}
+     *
+     * To be removed in 4.0
      */
     public function initRuntime(\Twig_Environment $environment)
     {
-        $this->renderer->setEnvironment($environment);
+        if (null !== $this->renderer) {
+            $this->renderer->setEnvironment($environment);
+        }
     }
 
     /**
@@ -69,7 +70,7 @@ class FormExtension extends \Twig_Extension implements \Twig_Extension_InitRunti
             new \Twig_SimpleFunction('form', null, array('node_class' => 'Symfony\Bridge\Twig\Node\RenderBlockNode', 'is_safe' => array('html'))),
             new \Twig_SimpleFunction('form_start', null, array('node_class' => 'Symfony\Bridge\Twig\Node\RenderBlockNode', 'is_safe' => array('html'))),
             new \Twig_SimpleFunction('form_end', null, array('node_class' => 'Symfony\Bridge\Twig\Node\RenderBlockNode', 'is_safe' => array('html'))),
-            new \Twig_SimpleFunction('csrf_token', array($this, 'renderCsrfToken')),
+            new \Twig_SimpleFunction('csrf_token', array('Symfony\Bridge\Twig\Form\TwigRenderer', 'renderCsrfToken')),
         );
     }
 
@@ -79,7 +80,7 @@ class FormExtension extends \Twig_Extension implements \Twig_Extension_InitRunti
     public function getFilters()
     {
         return array(
-            new \Twig_SimpleFilter('humanize', array($this, 'humanize')),
+            new \Twig_SimpleFilter('humanize', array('Symfony\Bridge\Twig\Form\TwigRenderer', 'humanize')),
         );
     }
 
@@ -89,60 +90,8 @@ class FormExtension extends \Twig_Extension implements \Twig_Extension_InitRunti
     public function getTests()
     {
         return array(
-            new \Twig_SimpleTest('selectedchoice', array($this, 'isSelectedChoice')),
+            new \Twig_SimpleTest('selectedchoice', 'Symfony\Bridge\Twig\Extension\twig_is_selected_choice'),
         );
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function renderCsrfToken($tokenId)
-    {
-        return $this->renderer->renderCsrfToken($tokenId);
-    }
-
-    /**
-     * Makes a technical name human readable.
-     *
-     * @param string $text The text to humanize
-     *
-     * @return string The humanized text
-     */
-    public function humanize($text)
-    {
-        return $this->renderer->humanize($text);
-    }
-
-    /**
-     * Returns whether a choice is selected for a given form value.
-     *
-     * Unfortunately Twig does not support an efficient way to execute the
-     * "is_selected" closure passed to the template by ChoiceType. It is faster
-     * to implement the logic here (around 65ms for a specific form).
-     *
-     * Directly implementing the logic here is also faster than doing so in
-     * ChoiceView (around 30ms).
-     *
-     * The worst option tested so far is to implement the logic in ChoiceView
-     * and access the ChoiceView method directly in the template. Doing so is
-     * around 220ms slower than doing the method call here in the filter. Twig
-     * seems to be much more efficient at executing filters than at executing
-     * methods of an object.
-     *
-     * @param ChoiceView   $choice        The choice to check
-     * @param string|array $selectedValue The selected value to compare
-     *
-     * @return bool Whether the choice is selected
-     *
-     * @see ChoiceView::isSelected()
-     */
-    public function isSelectedChoice(ChoiceView $choice, $selectedValue)
-    {
-        if (is_array($selectedValue)) {
-            return in_array($choice->value, $selectedValue, true);
-        }
-
-        return $choice->value === $selectedValue;
     }
 
     /**
@@ -152,4 +101,24 @@ class FormExtension extends \Twig_Extension implements \Twig_Extension_InitRunti
     {
         return 'form';
     }
+}
+
+/**
+ * Returns whether a choice is selected for a given form value.
+ *
+ * This is a function and not callable due to performance reasons.
+ *
+ * @param string|array $selectedValue The selected value to compare
+ *
+ * @return bool Whether the choice is selected
+ *
+ * @see ChoiceView::isSelected()
+ */
+function twig_is_selected_choice(ChoiceView $choice, $selectedValue)
+{
+    if (is_array($selectedValue)) {
+        return in_array($choice->value, $selectedValue, true);
+    }
+
+    return $choice->value === $selectedValue;
 }